# swagger.yaml
...
swagger: "2.0"
info:
    description: |
        member api
        [http://mbrtest.zoomye.com.tw](http://mbrtest.zoomye.com.tw)
    version: 1.0.0
    title: member api
host: mbrtest.zoomye.com.tw
basePath: /api/v1
...
schemes:
    - https
    - http
paths:
...
    /chops/consume:
        post:
            tags:
                - chops
            summary: 兌點
            operationId: consumeChops
            consumes:
                - application/json
            produces:
                - application/json
            parameters:
                - in: body
                  name: body
                  required: true
                  schema:
                      $ref: "#/definitions/ConsumeChopRequest"
            responses:
                "200":
                    description: Successfully consume chops
                "400":
                    description: Invalid input
    /chops/consume/{id}/void:
        post:
            tags:
                - chops
            summary: 取消兌點
            operationId: voidConsumeChops
            consumes:
                - application/json
            produces:
                - application/json
            parameters:
                - in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: body
                  required: true
                  schema:
                      $ref: "#/definitions/Member"
            responses:
                "200":
                    description: Successfully void consume chops
                    schema:
                        $ref: "#/definitions/ConsumeChopRule"
                "400":
                    description: Invalid input
...
definitions:
...
    Member:
        type: object
        required:
            - phone
            - password
            - name
            - url
            - serviceName
            - domain
        properties:
            id:
                type: integer
                format: int64
            first_name:
                type: string
            last_name:
                type: string
            phone:
                type: string
            password:
                type: string
            gender:
                type: string
                enum: [male, female, others, unknown]
            email:
                type: string
            address:
                type: string
            rank:
              $ref: "#/definitions/Rank"
    ConsumeChopRequest:
        type: object
        properties:
            member_id:
                type: string
            branch_id:
                type: string
            chops:
                type: integer
    ConsumeChopRule:
        type: object
        properties:
            name:
                type: string
            description:
                type: string
            payment_type:
                type: string
            type:
                type: string
            chops_per_unit:
                type: string
            unit_per_amount:
                type: string
            consume_max_percentage:
                type: string
            earn_chops_after_consume:
                type: boolean
            activated_at:
                type: string
                format: date
            expired_at:
                type: string
                format: date
            rank:
              $ref: "#/definitions/Rank"
...
    Rank:
        type: object
        properties:
            rank:
                type: integer
            name:
                type: string
...